Add scaffold class for API and route decorators
- Remove decorators module
- Remove class decorator for route decorators
- Update the API documentation
