+++ docs/api/app.md
@@ -1,3 +1,7 @@
 # `APIFlask`
 
 ::: apiflask.APIFlask
+
+## `APIScaffold`
+
+::: apiflask.scaffold.APIScaffold
