+++ docs/api/blueprint.md
@@ -1,3 +1,7 @@
 # `APIBlueprint`
 
 ::: apiflask.APIBlueprint
+
+## `APIScaffold`
+
+::: apiflask.scaffold.APIScaffold
