+++ mkdocs.yml
@@ -47,7 +47,6 @@ nav:
   - API Reference:
     - APIFlask: api/app.md
     - APIBlueprint: api/blueprint.md
-    - Decorators: api/decorators.md
     - Exceptions: api/exceptions.md
     - OpenAPI: api/openapi.md
     - Schemas: api/schemas.md
