+++ src/apiflask/route.py
@@ -94,73 +94,3 @@ def route_patch(cls):
 
     cls.add_url_rule = add_url_rule
     return cls
-
-
-def route_shortcuts(cls):
-    """A decorator to add route shortcuts for `Flask` and `Blueprint` objects.
-
-    Includes `get()`, `post()`, `put()`, `patch()`, `delete()`.
-
-    Examples:
-
-    ```python
-    from flask import Flask
-
-    @route_shortcuts
-    class APIFlask(Flask):
-        pass
-
-    app = APIFlask(__name__)
-
-    @app.get('/')
-    def hello():
-        return 'Hello!'
-    ```
-
-    *Version changed: 0.3.0*
-
-    - Turn base class into a class decorator.
-
-    *Version added: 0.2.0*
-    """
-    cls_route = cls.route
-
-    def _method_route(self, method: str, rule: str, options: t.Any):
-        if 'methods' in options:
-            raise RuntimeError('Use the "route" decorator to use the "methods" argument.')
-
-        def decorator(f):
-            if isinstance(f, MethodViewType):
-                raise RuntimeError(
-                    'The route shortcuts cannot be used with "MethodView" classes, '
-                    'use the "route" decorator instead.'
-                )
-            return cls_route(self, rule, methods=[method], **options)(f)
-        return decorator
-
-    def get(self, rule: str, **options: t.Any):
-        """Shortcut for `app.route()`."""
-        return _method_route(self, 'GET', rule, options)
-
-    def post(self, rule: str, **options: t.Any):
-        """Shortcut for `app.route(methods=['POST'])`."""
-        return _method_route(self, 'POST', rule, options)
-
-    def put(self, rule: str, **options: t.Any):
-        """Shortcut for `app.route(methods=['PUT'])`."""
-        return _method_route(self, 'PUT', rule, options)
-
-    def patch(self, rule: str, **options: t.Any):
-        """Shortcut for `app.route(methods=['PATCH'])`."""
-        return _method_route(self, 'PATCH', rule, options)
-
-    def delete(self, rule: str, **options: t.Any):
-        """Shortcut for `app.route(methods=['DELETE'])`."""
-        return _method_route(self, 'DELETE', rule, options)
-
-    cls.get = get
-    cls.post = post
-    cls.put = put
-    cls.patch = patch
-    cls.delete = delete
-    return cls
